+++ src/Console/MessageTrait.php
@@ -44,9 +44,9 @@ trait MessageTrait
     /**
      * Get the value of message
      *
-     * @return  string
+     * @return string
      */
-    public function getMessage()
+    public function getMessage(): string
     {
         return self::$message;
     }
@@ -65,6 +65,7 @@ trait MessageTrait
     }
 
     /**
+     * @param string $message
      * @param bool $space
      * 
      * @return self
@@ -85,6 +86,7 @@ trait MessageTrait
     }
 
     /**
+     * @param string $message
      * @param bool $space
      * 
      * @return self
@@ -105,6 +107,7 @@ trait MessageTrait
     }
 
     /**
+     * @param string $message
      * @param bool $space
      * 
      * @return self
@@ -125,6 +128,7 @@ trait MessageTrait
     }
 
     /**
+     * @param string $message
      * @param bool $space
      * 
      * @return self
@@ -145,6 +149,7 @@ trait MessageTrait
     }
 
     /**
+     * @param string $message
      * @param bool $space
      * 
      * @return self
@@ -175,6 +180,8 @@ trait MessageTrait
     }
 
     /**
+     * @param bool $repeat
+     * 
      * @return self
      */
     public function break($repeat = false): self
@@ -214,9 +221,9 @@ trait MessageTrait
     }
 
     /**
-     * @return bool
+     * @return mixed
      */
-    public function colorIsSupported()
+    public function colorIsSupported(): mixed
     {
         if (DIRECTORY_SEPARATOR === '\\') {
             if (function_exists('sapi_windows_vt100_support') && @sapi_windows_vt100_support(STDOUT)) {
@@ -231,9 +238,9 @@ trait MessageTrait
     }
 
     /**
-     * @return bool
+     * @return mixed
      */
-    public function are256ColorsSupported()
+    public function are256ColorsSupported(): mixed
     {
         if (DIRECTORY_SEPARATOR === '\\') {
             return function_exists('sapi_windows_vt100_support') && @sapi_windows_vt100_support(STDOUT);