+++ src/Interface/SolutionInterface.php
@@ -6,5 +6,8 @@ use ModernPHPException\Solution;
 
 interface SolutionInterface
 {
+    /**
+     * @return Solution
+     */
     public function getSolution(): Solution;
 }