+++ src/Trait/RenderTrait.php
@@ -72,6 +72,11 @@ trait RenderTrait
 
         if ($this->format == "json" || $this->config['type'] === "json") {
             $this->renderJson();
+        } else if ($this->format == "text" || $this->config['type'] === "text") {
+            //$this->renderText();
+
+            include_once dirname(__DIR__) . '/View/templates/text-error.php';
+            exit;
         }
 
         if ($this->config['title'] !== "") {
@@ -117,6 +122,37 @@ trait RenderTrait
     /**
      * @return void
      */
+    private function renderText(): void
+    {
+        if (isset($this->info_error_exception['type_exception'])) {
+            echo "[" . $this->info_error_exception['type_exception'] . "] " . $this->info_error_exception['message'] . '<br>';
+            //$this->renderSolution();
+            
+            echo "File: " . $this->info_error_exception['file'] . '<br>';
+            echo "Line: " . $this->info_error_exception['line'] . '<br>';
+            //$this->getLines($this->info_error_exception['file'], $this->info_error_exception['line']);
+
+            /* if (!empty($this->trace)) {
+                echo PHP_EOL;
+                $this->warning(count($this->trace) . " other error(s)")->print()->break();
+            }
+
+            foreach ($this->trace as $trace) {
+                $this->info("Line: " . $trace['line'] . " | File: " . $trace['file'])->print()->break();
+            } */
+        } else {
+            echo "[" . $this->getError() . "] " . $this->info_error_exception['message'] . '<br>';
+            echo "File: " . $this->info_error_exception['file'] . '<br>';
+            echo "Line: " . $this->info_error_exception['line'] . '<br>';
+            //$this->getLines($this->info_error_exception['file'], $this->info_error_exception['line']);
+        }
+
+        exit;
+    }
+
+    /**
+     * @return void
+     */
     private function renderCli(): void
     {
         $verify = $this->isCli();
@@ -150,7 +186,13 @@ trait RenderTrait
         }
     }
 
-    private function getLines(string $context, int $line)
+    /**
+     * @param string $context
+     * @param int $line
+     * 
+     * @return mixed
+     */
+    private function getLines(string $context, int $line): mixed
     {
         for ($i = 0; $i < 7; $i++) {
             $lines_up[] = $line + $i;
@@ -215,6 +257,8 @@ trait RenderTrait
         echo PHP_EOL . PHP_EOL;
 
         $this->line("------------------------------------------------------------------------")->print()->break();
+
+        return $this;
     }
 
     /**